Move vallox coordinator to separate module
