+++ homeassistant/components/vallox/sensor.py
@@ -24,7 +24,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
 from homeassistant.util import dt as dt_util
 
-from . import ValloxDataUpdateCoordinator, ValloxEntity
+from . import ValloxEntity
 from .const import (
     DOMAIN,
     METRIC_KEY_MODE,
@@ -32,6 +32,7 @@ from .const import (
     VALLOX_CELL_STATE_TO_STR,
     VALLOX_PROFILE_TO_PRESET_MODE_REPORTABLE,
 )
+from .coordinator import ValloxDataUpdateCoordinator
 
 
 class ValloxSensorEntity(ValloxEntity, SensorEntity):
